Fix unit tests
(cherry picked from commit 42db96f8db140fd550139ac6e5c6c9bb648e8b84)
