+++ crates/light-client-prover/src/lib.rs
@@ -16,3 +16,7 @@ pub mod runner;
 mod services;
 #[cfg(test)]
 mod tests;
+
+/// Test utilities
+#[cfg(test)]
+pub use tests::test_utils::*;
