+++ crates/light-client-prover/src/tests/test_utils.rs
@@ -2,7 +2,9 @@ use std::collections::BTreeMap;
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use alloy_primitives::eip191_hash_message;
+use alloy_primitives::{eip191_hash_message, B256};
+use alloy_signer::SignerSync;
+use alloy_signer_local::PrivateKeySigner;
 use rand::{thread_rng, Rng};
 use sov_mock_da::{MockAddress, MockBlob, MockDaSpec, MockDaVerifier};
 use sov_mock_zkvm::{MockCodeCommitment, MockJournal, MockProof, MockZkvm};
@@ -21,6 +23,7 @@ use sov_rollup_interface::zk::light_client_proof::output::LightClientCircuitOutp
 use crate::circuit::accessors::ChunkAccessor;
 use crate::circuit::LightClientProofCircuit;
 
+/// Test private keys used for generating signatures in tests
 pub const TEST_PRIVATE_KEYS: [&str; 5] = [
     "79122E48DF1A002FB6584B2E94D0D50F95037416C82DAF280F21CD67D17D9077",
     "79122E48DF1A002FB6584B2E94D0D50F95037416C82DAF280F21CD67D17D9076",
@@ -175,13 +178,6 @@ pub(crate) fn create_serialized_mock_proof(
 
     mock_proof.encode_to_vec()
 }
-pub(crate) fn from_vec_to_sigs(vec: Vec<Vec<u8>>) -> [[u8; 64]; 5] {
-    let mut sigs = Vec::new();
-    for v in vec.into_iter() {
-        sigs.push(v.try_into().unwrap());
-    }
-    sigs.try_into().unwrap()
-}
 
 pub(crate) fn create_prev_lcp_serialized(
     output: LightClientCircuitOutput,
@@ -201,18 +197,66 @@ pub(crate) fn create_prev_lcp_serialized(
     mock_proof.encode_to_vec()
 }
 
-fn eip191_sign(msg: &[u8], secret_key: &[u8; 32]) -> (k256::ecdsa::Signature, [u8; 32]) {
-    use alloy_signer::SignerSync;
-    use alloy_signer_local::PrivateKeySigner;
+/// Converts a vector of signatures in Vec<u8> format to an array of signatures in [u8; 64] format
+pub fn from_vec_to_sigs(vec: Vec<(Vec<u8>, u8)>) -> [([u8; 64], u8); 3] {
+    let mut sigs = Vec::new();
+    for (v, i) in vec.into_iter() {
+        sigs.push((v.try_into().unwrap(), i));
+    }
+    sigs.try_into().unwrap()
+}
 
-    let signer = PrivateKeySigner::from_bytes(&secret_key.into()).unwrap();
+/// Generates 5 valid keypairs and returns the public keys and signers from the given private keys
+pub fn generate_initial_pub_keys_with_signers_from_pks(
+    private_keys: [[u8; 32]; 5],
+) -> ([[u8; 33]; 5], Vec<PrivateKeySigner>) {
+    let mut initial_da_pubkeys = [[0u8; 33]; 5];
+    let mut signers = Vec::new();
 
-    let prehash = eip191_hash_message(msg);
+    // Generate 5 valid keypairs and signatures
+    for (i, secret_key) in private_keys.iter().enumerate() {
+        let signer = PrivateKeySigner::from_bytes(&secret_key.into()).unwrap();
+        let verifying_key = signer.credential().verifying_key();
+        let pubkey = verifying_key.to_sec1_bytes();
+        initial_da_pubkeys[i] = pubkey.to_vec().try_into().unwrap();
+        signers.push(signer);
+    }
 
-    let sig = signer.sign_hash_sync(&prehash).unwrap();
-    let signature = k256::ecdsa::Signature::from_slice(&sig.as_bytes()[0..64]).unwrap();
+    (initial_da_pubkeys, signers)
+}
 
-    (signature, *prehash)
+/// Generates 5 valid keypairs and returns the public keys and signers
+pub fn generate_initial_pub_keys_with_signers() -> ([[u8; 33]; 5], Vec<PrivateKeySigner>) {
+    let mut initial_da_pubkeys = [[0u8; 33]; 5];
+    let mut signers = Vec::new();
+
+    // Generate 5 valid keypairs and signatures
+    for i in 0..5 {
+        let secret_key = [i as u8 + 1; 32];
+        let signer = PrivateKeySigner::from_bytes(&secret_key.into()).unwrap();
+        let verifying_key = signer.credential().verifying_key();
+        let pubkey = verifying_key.to_sec1_bytes();
+        initial_da_pubkeys[i] = pubkey.to_vec().try_into().unwrap();
+        signers.push(signer);
+    }
+
+    (initial_da_pubkeys, signers)
+}
+
+/// Creates 3 valid signatures from the first 3 signers for the given prehash
+pub fn create_valid_signatures(
+    signers: &Vec<PrivateKeySigner>,
+    prehash: &B256,
+) -> [([u8; 64], u8); 3] {
+    let mut signatures_in_inscription = Vec::new();
+
+    for (i, signer) in signers.iter().enumerate().take(3) {
+        let sig = signer.sign_hash_sync(prehash).unwrap();
+        let signature = sig.as_bytes()[0..64].to_vec();
+        signatures_in_inscription.push((signature, i as u8));
+    }
+
+    from_vec_to_sigs(signatures_in_inscription)
 }
 
 pub(crate) fn create_new_method_id_tx(
@@ -229,18 +273,16 @@ pub(crate) fn create_new_method_id_tx(
     })
     .unwrap();
 
-    let mut signatures = vec![];
-    for pk_bytes in pk_bytes_arr {
-        let (sig, _hash) = eip191_sign(msg.as_slice(), &pk_bytes);
-        signatures.push(sig.to_vec());
-    }
+    let (_initial_pubkeys, signers) = generate_initial_pub_keys_with_signers_from_pks(pk_bytes_arr);
+    let prehash = eip191_hash_message(&msg);
+    let signatures_with_index = create_valid_signatures(&signers, &prehash);
 
     let da_data = DataOnDa::BatchProofMethodId(BatchProofMethodId {
         body: BatchProofMethodIdBody {
             method_id: new_method_id,
             activation_l2_height: activation_height,
         },
-        signatures: from_vec_to_sigs(signatures),
+        signatures_with_index,
     });
 
     let da_data_ser = borsh::to_vec(&da_data).expect("should serialize");
